@@ -1,18 +1,16 @@
 name: fluttervimeoplayer
 description: A flutter package wrapping the vimeo player
-version: 0.0.1
-author:
-homepage:
+version: 0.0.2
+homepage: https://github.com/ThuAbLKA/flutter-vimeo-player.git
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_inappwebview: ^3.4.0+2
+  flutter_inappwebview: ^5.3.2
   async: ^2.4.0
-  positioned_tap_detector: ^1.0.3
 
 dev_dependencies:
   flutter_test:
@@ -23,7 +21,7 @@ dev_dependencies:
 
 # The following section is specific to Flutter.
 flutter:
-
+  
   # To add assets to your package, add an assets section, like this:
   # assets:
   #  - images/a_dot_burr.jpeg
