@@ -1,6 +1,4 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:fluttervimeoplayer/flutter_vimeo_player.dart';
 
-
-void main() {
-}
+void main() {}
